# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "agent-toolbox.ai API",
    "version": "0.1.0",
    "description": "The quality layer for AI agents — deterministic, offline pre-action gates. Each tool returns a PASS/FLAG/BLOCK verdict plus a tamper-evident SHA-256 certificate."
  },
  "servers": [
    {
      "url": "https://api.agent-toolbox.ai",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Solana transaction signature or API key as a Bearer token. Omit entirely to use the free tier (10 calls per IP, no auth)."
      }
    },
    "schemas": {
...
      "FinanceUnitsRequest": {
        "type": "object",
        "properties": {
          "tokenAddress": {
            "type": "string"
          },
          "rawAmount": {
            "type": "string"
          },
          "uiAmount": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "chain": {
            "type": "string",
            "enum": [
              "solana",
              "ethereum",
              "bsc",
              "polygon",
              "base",
              "arbitrum"
            ]
          }
        },
        "required": [
          "tokenAddress",
          "rawAmount",
          "uiAmount",
          "chain"
        ]
      },
      "FinanceUnitsResponse": {
        "type": "object",
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "FLAG",
              "BLOCK"
            ]
          },
          "score": {
            "type": "number"
          },
          "risks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "severity": {
                  "type": "string",
                  "enum": [
                    "info",
                    "warn",
                    "critical"
                  ]
                },
                "detail": {
                  "type": "string"
                }
              },
              "required": [
                "type",
                "severity",
                "detail"
              ]
            }
          },
          "latencyMs": {
            "type": "number",
            "description": "Server compute time in ms."
          },
          "authoritative_decimals": {
            "type": [
              "integer",
              "null"
            ]
          },
          "expected_raw": {
            "type": [
              "string",
              "null"
            ]
          },
          "actual_raw": {
            "type": "string"
          },
          "deviation_pct": {
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "verdict",
          "score",
          "risks",
          "latencyMs",
          "authoritative_decimals",
          "expected_raw",
          "actual_raw",
          "deviation_pct"
        ]
      },
      "FinancePriceRequest": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string"
          },
          "tokenAddress": {
            "type": "string"
          },
          "assetType": {
            "type": "string",
            "enum": [
              "crypto",
              "stock",
              "forex"
            ]
          },
          "proposedPrice": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "maxAgeSeconds": {
            "type": "integer",
            "exclusiveMinimum": 0
          },
          "divergenceThresholdPct": {
            "type": "number",
            "exclusiveMinimum": 0
          }
        },
        "required": [
          "assetType"
        ]
      },
      "FinancePriceResponse": {
        "type": "object",
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "FLAG",
              "BLOCK"
            ]
          },
          "score": {
            "type": "number"
          },
          "risks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "severity": {
                  "type": "string",
                  "enum": [
                    "info",
                    "warn",
                    "critical"
                  ]
                },
                "detail": {
                  "type": "string"
                }
              },
              "required": [
                "type",
                "severity",
                "detail"
              ]
            }
          },
          "latencyMs": {
            "type": "number",
            "description": "Server compute time in ms."
          },
          "sources": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "priceUsd": {
                  "type": "number"
                },
                "ageSeconds": {
                  "type": [
                    "number",
                    "null"
                  ]
                },
                "available": {
                  "type": "boolean"
                }
              },
              "required": [
                "name",
                "priceUsd",
                "ageSeconds",
                "available"
              ]
            }
          },
          "consensusPrice": {
            "type": [
              "number",
              "null"
            ]
          },
          "proposedPriceDeviation": {
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "verdict",
          "score",
          "risks",
          "latencyMs",
          "sources",
          "consensusPrice",
          "proposedPriceDeviation"
        ]
      },
      "FinanceSymbolRequest": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string"
          },
          "assetType": {
            "type": "string",
            "enum": [
              "crypto",
              "stock"
            ]
          },
          "expectedName": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "assetType"
        ]
      },
      "FinanceSymbolResponse": {
        "type": "object",
        "properties": {
          "found": {
            "type": "boolean"
          },
          "matches": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "symbol": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "exchange": {
                  "type": "string"
                },
                "liquidity": {
                  "type": "number"
                }
              },
              "required": [
                "symbol",
                "name"
              ]
            }
          },
          "ambiguous": {
            "type": "boolean"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "FLAG",
              "BLOCK"
            ]
          }
        },
        "required": [
          "found",
          "matches",
          "ambiguous",
          "verdict"
        ]
      },
      "FinanceTokenRiskRequest": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "chain": {
            "type": "string",
            "enum": [
              "solana",
              "ethereum",
              "bsc",
              "polygon",
              "base",
              "arbitrum"
            ]
          },
          "maxRugScore": {
            "type": "number"
          },
          "requireLpLocked": {
            "type": "boolean"
          },
          "blockIfMintAuthority": {
            "type": "boolean"
          },
          "blockIfFreezeAuthority": {
            "type": "boolean"
          }
        },
        "required": [
          "address",
          "chain"
        ]
      },
      "FinanceTokenRiskResponse": {
        "type": "object",
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "FLAG",
              "BLOCK"
            ]
          },
          "score": {
            "type": "number"
          },
          "risks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "severity": {
                  "type": "string",
                  "enum": [
                    "info",
                    "warn",
                    "critical"
                  ]
                },
                "detail": {
                  "type": "string"
                }
              },
              "required": [
                "type",
                "severity",
                "detail"
              ]
            }
          },
          "latencyMs": {
            "type": "number",
            "description": "Server compute time in ms."
          },
          "rugScore": {
            "type": [
              "number",
              "null"
            ]
          },
          "mintAuthorityActive": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "freezeAuthorityActive": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "lpLockedPct": {
            "type": [
              "number",
              "null"
            ]
          },
          "specificRisks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "verdict",
          "score",
          "risks",
          "latencyMs",
          "rugScore",
          "mintAuthorityActive",
          "freezeAuthorityActive",
          "lpLockedPct",
          "specificRisks"
        ]
      },
      "FinanceSlippageRequest": {
        "type": "object",
        "properties": {
          "tokenAddress": {
            "type": "string"
          },
          "chain": {
            "type": "string"
          },
          "tradeUsd": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "maxPriceImpactPct": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "minLiquidityUsd": {
            "type": "number",
            "exclusiveMinimum": 0
          }
        },
        "required": [
          "tokenAddress",
          "chain",
          "tradeUsd"
        ]
      },
      "FinanceSlippageResponse": {
        "type": "object",
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "FLAG",
              "BLOCK"
            ]
          },
          "score": {
            "type": "number"
          },
          "risks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string"
                },
                "severity": {
                  "type": "string",
                  "enum": [
                    "info",
                    "warn",
                    "critical"
                  ]
                },
                "detail": {
                  "type": "string"
                }
              },
              "required": [
                "type",
                "severity",
                "detail"
              ]
            }
          },
          "latencyMs": {
            "type": "number",
            "description": "Server compute time in ms."
          },
          "poolLiquidityUsd": {
            "type": [
              "number",
              "null"
            ]
          },
          "estimatedPriceImpactPct": {
            "type": [
              "number",
              "null"
            ]
          },
          "volume24h": {
            "type": [
              "number",
              "null"
            ]
          },
          "washTradingFlag": {
            "type": "boolean"
          }
        },
        "required": [
          "verdict",
          "score",
          "risks",
          "latencyMs",
          "poolLiquidityUsd",
          "estimatedPriceImpactPct",
          "volume24h",
          "washTradingFlag"
        ]
      },
      "FinanceOrderRiskRequest": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string"
          },
          "tokenAddress": {
            "type": "string"
          },
          "assetType": {
            "type": "string",
            "enum": [
              "crypto",
              "stock"
            ]
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ]
          },
          "tradeUsd": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "portfolioValueUsd": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "chain": {
            "type": "string"
          },
          "leverage": {
            "type": "number",
            "exclusiveMinimum": 0
          }
        },
        "required": [
          "assetType",
          "side",
          "tradeUsd"
        ]
      },
      "FinanceOrderRiskResponse": {
        "type": "object",
        "properties": {
          "verdict": {
            "type": "string",
            "enum": [
              "PASS",
              "FLAG",
              "BLOCK"
            ]
          },
          "overallScore": {
            "type": "number"
          },
          "checks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "verdict": {
                  "type": "string",
                  "enum": [
                    "PASS",
                    "FLAG",
                    "BLOCK"
                  ]
                },
                "score": {
                  "type": "number"
                },
                "risks": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string"
                      },
                      "severity": {
                        "type": "string",
                        "enum": [
                          "info",
                          "warn",
                          "critical"
                        ]
                      },
                      "detail": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "type",
                      "severity",
                      "detail"
                    ]
                  }
                },
                "latencyMs": {
                  "type": "number",
                  "description": "Server compute time in ms."
                },
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "verdict",
                "score",
                "risks",
                "latencyMs",
                "name"
              ]
            }
          },
          "blockedBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "latencyMs": {
            "type": "number",
            "description": "Server compute time in ms."
          }
        },
        "required": [
          "verdict",
          "overallScore",
          "checks",
          "blockedBy",
          "latencyMs"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "docs": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/v1/finance/units": {
      "post": {
        "operationId": "financeUnits",
        "summary": "Guard against decimal-scaling errors by verifying raw on-chain amount against the token's authoritative decimals.",
        "description": "Cost: 1 credit(s). Free tier: 10 calls per IP (no auth).",
        "tags": [
          "finance"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinanceUnitsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceUnitsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Free tier exhausted or insufficient credits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/finance/price": {
      "post": {
        "operationId": "financePrice",
        "summary": "Cross-validate an asset price against multiple sources (CoinGecko, DexScreener, Yahoo Finance) and flag stale or hallucinated prices.",
        "description": "Cost: 2 credit(s). Free tier: 10 calls per IP (no auth).",
        "tags": [
          "finance"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancePriceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancePriceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Free tier exhausted or insufficient credits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/finance/symbol": {
      "post": {
        "operationId": "financeSymbol",
        "summary": "Resolve a ticker symbol or token to a confirmed on-chain identity, detecting ambiguous or imposter tokens.",
        "description": "Cost: 1 credit(s). Free tier: 10 calls per IP (no auth).",
        "tags": [
          "finance"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinanceSymbolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceSymbolResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Free tier exhausted or insufficient credits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/finance/token/risk": {
      "post": {
        "operationId": "financeTokenRisk",
        "summary": "Run a rug-pull risk assessment on a token using RugCheck — checks mint/freeze authority, LP lock, and overall rug score.",
        "description": "Cost: 3 credit(s). Free tier: 10 calls per IP (no auth).",
        "tags": [
          "finance"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinanceTokenRiskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceTokenRiskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Free tier exhausted or insufficient credits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/finance/slippage": {
      "post": {
        "operationId": "financeSlippage",
        "summary": "Check pool liquidity and estimate price impact (slippage) for a proposed trade size.",
        "description": "Cost: 2 credit(s). Free tier: 10 calls per IP (no auth).",
        "tags": [
          "finance"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinanceSlippageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceSlippageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Free tier exhausted or insufficient credits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/finance/order/risk": {
      "post": {
        "operationId": "financeOrderRisk",
        "summary": "Full pre-trade risk gate — runs rug, liquidity, price, and position checks in parallel and returns a composite verdict.",
        "description": "Cost: 5 credit(s). Free tier: 10 calls per IP (no auth).",
        "tags": [
          "finance"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinanceOrderRiskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceOrderRiskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Free tier exhausted or insufficient credits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/finance/position/check": {
      "post": {
        "operationId": "financePositionCheck",
        "summary": "Deterministic position-limit and guardian-rules gate — enforces kill switch, daily-loss cap, position size, leverage, and allowed-asset constraints.",
        "description": "Cost: 1 credit(s). Free tier: 10 calls per IP (no auth).",
        "tags": [
          "finance"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancePositionCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancePositionCheckResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Free tier exhausted or insufficient credits.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}